+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+- Implement `serde::Serialize` and `serde::Deserialize` for `AccountSignature`, `CredentialSignature`, `Signature`, `SignatureEcdsaSecp256k1` and `SignatureEd25519` when feature `derive-serde` is enabled.
+- Set minimum supported Rust version to 1.73.
+- Change `HashBytes::new` is now a `const` function.
+
 ## concordium-contracts-common 9.0.0 (2024-01-22)
 
 - Add `SchemaType` implementation for `&str`.